+++ examples/transit/index.php
@@ -5,7 +5,7 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1.0,user-scalable=no">
-    <title>TSEP Checkout Examples</title>
+    <title>TransIT Examples</title>
   </head>
   <body>
     <h1>Examples</h1>
