+++ src/internal/gateways/transit/action-normalize-response.ts
