+++ src/internal/gateways/transit/action-validate-data.ts
